// managme/src/models/story.ts
export interface Story {
  id: string;
  name: string;
  description: string;
  priority: StoryPriority;
  projectId: string;
  createdAt: string;
  status: StoryStatus;
  ownerId: string;
}

export type StoryPriority = "low" | "medium" | "high";

export type StoryStatus = "todo" | "doing" | "done";